backend/requirements.txt
Django>=6.0,<6.1
psycopg[binary]>=3.2,<4.0
Pillow>=10.0,<12.0
python-dotenv>=1.0,<2.0
gunicorn>=22.0,<23.0
python-slugify>=8.0,<9.0
